Fix formatting and style issues
