+++ conf/broker.conf
@@ -295,28 +295,31 @@ keepAliveIntervalSeconds=30
 brokerServicePurgeInactiveFrequencyInSeconds=60
 
 # Number of samples to use for short term time window
-numShortSamples=50
+loadManagerNumberOfSamplesShortTermWindow=50
 
 # Number of samples to use for long term time window
-numLongSamples=1000
+loadManagerNumberOfSamplesLongTermWindow=1000
 
 # How often in seconds to update the broker data
-brokerDataUpdateIntervalSeconds=60
+loadManagerBrokerDataUpdateIntervalInSeconds=60
 
 # How often in seconds to update the bundle data
-bundleDataUpdateIntervalSeconds=60
+loadManagerBundleDataUpdateIntervalInSeconds=60
 
 # Default throughput in to assume for new bundles
-defaultMsgThroughputIn=50000
+loadManagerDefaultMessageThroughputIn=50000
 
 # Default throughput out to assume for new bundles
-defaultMsgThroughputOut=50000
+loadManagerDefaultMessageThroughputOut=50000
 
 # Default message rate in to assume for new bundles
-defaultMsgRateIn=50
+loadManagerDefaultMessageRateIn=50
 
 # Default message rate out to assume for new bundles
-defaultMsgRateOut=50
+loadManagerDefaultMessageRateOut=50
 
 # Name of load manager to use
-loadManagerName=SimpleLoadManager
+loadManagerClassName=com.yahoo.pulsar.broker.loadbalance.impl.SimpleLoadManagerImpl
+
+# Name of placement strategy to use for new load manager API
+loadManagerPlacementStrategyClassName=com.yahoo.pulsar.broker.loadbalance.impl.LeastLongTermMessageRate
