+++ pulsar-broker/src/main/java/com/yahoo/pulsar/broker/loadbalance/LoadManager.java
@@ -20,91 +20,103 @@ import java.util.List;
 import com.yahoo.pulsar.broker.PulsarServerException;
 import com.yahoo.pulsar.broker.PulsarService;
 import com.yahoo.pulsar.broker.ServiceConfiguration;
-import com.yahoo.pulsar.broker.loadbalance.impl.NewLoadManagerImpl;
-import com.yahoo.pulsar.broker.loadbalance.impl.NewLoadManagerWrapper;
+import com.yahoo.pulsar.broker.loadbalance.impl.ModularLoadManagerWrapper;
 import com.yahoo.pulsar.broker.loadbalance.impl.SimpleLoadManagerImpl;
 import com.yahoo.pulsar.broker.stats.Metrics;
 import com.yahoo.pulsar.common.naming.ServiceUnitId;
 import com.yahoo.pulsar.common.policies.data.loadbalancer.LoadReport;
 
 /**
- * LoadManager runs though set of load reports collected from different brokers and generates a recommendation of
- * namespace/ServiceUnit placement on machines/ResourceUnit. Each Concrete Load Manager will use different algorithms to
- * generate this mapping.
+ * LoadManager runs though set of load reports collected from different brokers
+ * and generates a recommendation of namespace/ServiceUnit placement on
+ * machines/ResourceUnit. Each Concrete Load Manager will use different
+ * algorithms to generate this mapping.
  *
- * Concrete Load Manager is also return the least loaded broker that should own the new namespace.
+ * Concrete Load Manager is also return the least loaded broker that should own
+ * the new namespace.
  */
 public interface LoadManager {
 
-    public void start() throws PulsarServerException;
+	public void start() throws PulsarServerException;
 
-    /**
-     * Is centralized decision making to assign a new bundle.
-     */
-    boolean isCentralized();
+	/**
+	 * Is centralized decision making to assign a new bundle.
+	 */
+	boolean isCentralized();
 
-    /**
-     * Returns the Least Loaded Resource Unit decided by some algorithm or criteria which is implementation specific
-     */
-    ResourceUnit getLeastLoaded(ServiceUnitId su) throws Exception;
+	/**
+	 * Returns the Least Loaded Resource Unit decided by some algorithm or
+	 * criteria which is implementation specific
+	 */
+	ResourceUnit getLeastLoaded(ServiceUnitId su) throws Exception;
 
-    /**
-     * Generate the load report
-     */
-    LoadReport generateLoadReport() throws Exception;
+	/**
+	 * Generate the load report
+	 */
+	LoadReport generateLoadReport() throws Exception;
 
-    /**
-     * Set flag to force load report update
-     */
-    void setLoadReportForceUpdateFlag();
+	/**
+	 * Set flag to force load report update
+	 */
+	void setLoadReportForceUpdateFlag();
 
-    /**
-     * Publish the current load report on ZK
-     */
-    void writeLoadReportOnZookeeper() throws Exception;
+	/**
+	 * Publish the current load report on ZK
+	 */
+	void writeLoadReportOnZookeeper() throws Exception;
 
-    /**
-     * Update namespace bundle resource quota on ZK
-     */
-    void writeResourceQuotasToZooKeeper() throws Exception;
+	/**
+	 * Update namespace bundle resource quota on ZK
+	 */
+	void writeResourceQuotasToZooKeeper() throws Exception;
 
-    /**
-     * Generate load balancing stats metrics
-     */
-    List<Metrics> getLoadBalancingMetrics();
+	/**
+	 * Generate load balancing stats metrics
+	 */
+	List<Metrics> getLoadBalancingMetrics();
 
-    /**
-     * Unload a candidate service unit to balance the load
-     */
-    void doLoadShedding();
+	/**
+	 * Unload a candidate service unit to balance the load
+	 */
+	void doLoadShedding();
 
-    /**
-     * Namespace bundle split
-     */
-    void doNamespaceBundleSplit() throws Exception;
+	/**
+	 * Namespace bundle split
+	 */
+	void doNamespaceBundleSplit() throws Exception;
 
-    /**
-     * Determine the broker root.
-     */
-    String getBrokerRoot();
+	/**
+	 * Determine the broker root.
+	 */
+	String getBrokerRoot();
 
-    /**
-     * Removes visibility of current broker from loadbalancer list so, other brokers can't redirect any request to this
-     * broker and this broker won't accept new connection requests.
-     *
-     * @throws Exception
-     */
-    public void disableBroker() throws Exception;
+	/**
+	 * Removes visibility of current broker from loadbalancer list so, other
+	 * brokers can't redirect any request to this broker and this broker won't
+	 * accept new connection requests.
+	 *
+	 * @throws Exception
+	 */
+	public void disableBroker() throws Exception;
 
-    public void stop() throws PulsarServerException;
+	public void stop() throws PulsarServerException;
 
-    static LoadManager create(final ServiceConfiguration conf, final PulsarService pulsar) {
-        switch (conf.getLoadManagerName()) {
-            case "NewLoadManager":
-                return new NewLoadManagerWrapper(new NewLoadManagerImpl(pulsar));
-            case "SimpleLoadManager":
-            default:
-                return new SimpleLoadManagerImpl(pulsar);
-        }
-    }
+	static LoadManager create(final PulsarService pulsar) {
+		try {
+			final ServiceConfiguration conf = pulsar.getConfiguration();
+			final Class<?> loadManagerClass = Class.forName(conf.getLoadManagerClassName());
+			// Assume there is a constructor with one argument of PulsarService.
+			final Object loadManagerInstance = loadManagerClass.getConstructor(PulsarService.class).newInstance(pulsar);
+			if (loadManagerInstance instanceof LoadManager) {
+				return (LoadManager) loadManagerInstance;
+			} else if (loadManagerInstance instanceof ModularLoadManager) {
+				return new ModularLoadManagerWrapper((ModularLoadManager) loadManagerInstance);
+			}
+		} catch (Exception e) {
+			// Ignore
+		}
+		// If we failed to create a load manager, default to
+		// SimpleLoadManagerImpl.
+		return new SimpleLoadManagerImpl(pulsar);
+	}
 }
