+++ pulsar-common/src/main/java/com/yahoo/pulsar/common/policies/data/loadbalancer/ServiceLookupData.java
@@ -2,11 +2,11 @@ package com.yahoo.pulsar.common.policies.data.loadbalancer;
 
 // For backwards compatibility purposes.
 public interface ServiceLookupData {
-    public String getWebServiceUrl();
+	public String getWebServiceUrl();
 
-    public String getWebServiceUrlTls();
+	public String getWebServiceUrlTls();
 
-    public String getPulsarServiceUrl();
+	public String getPulsarServiceUrl();
 
-    public String getPulsarServiceUrlTls();
+	public String getPulsarServiceUrlTls();
 }
